requirements.txt
graphviz
matplotlib
Pillow
numpy
imageio